[111] - make lot profile page general, add general module
