+++ src/Auctionify.API/ClientApp/src/app/app-routing.module.ts
@@ -4,6 +4,8 @@ import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './components/home/home.component';
 import { isSellerGuard } from './guards/seller/is-seller.guard';
 import { isBuyerGuard } from './guards/buyer/is-buyer.guard';
+import { LotProfileComponent } from './components/general/lot-profile/lot-profile.component';
+import { isLoggedInGuard } from './guards/is-logged-in.guard';
 
 const routes: Routes = [
     { path: '', redirectTo: 'home', pathMatch: 'full' },
@@ -25,6 +27,11 @@ const routes: Routes = [
             ),
         data: { breadcrumb: { skip: true } },
         canActivate: [isBuyerGuard],
+    },
+    {
+        path: 'get-profile/:id',
+        component: LotProfileComponent,
+        canActivate: [isLoggedInGuard],
     },
 ];
