+++ src/Auctionify.API/ClientApp/src/app/app.module.ts
@@ -9,12 +9,12 @@ import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { AuthorizeInterceptor } from './api-authorization/authorize.interceptor';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { SellerModule } from './components/seller/seller.module';
-import { DashboardComponent } from './components/seller/dashboard/dashboard.component';
 import { LayoutModule } from './layout/layout.module';
 import { BuyerModule } from './components/buyer/buyer.module';
+import { GeneralModule } from './components/general/general.module';
 
 @NgModule({
-    declarations: [AppComponent, HomeComponent, DashboardComponent],
+    declarations: [AppComponent, HomeComponent],
     imports: [
         BrowserModule,
         AppRoutingModule,
@@ -24,6 +24,7 @@ import { BuyerModule } from './components/buyer/buyer.module';
         SellerModule,
         BuyerModule,
         LayoutModule,
+        GeneralModule,
     ],
 
     providers: [
