+++ src/Auctionify.API/ClientApp/src/app/components/general/general.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { LotProfileComponent } from './lot-profile/lot-profile.component';
+
+@NgModule({
+    declarations: [LotProfileComponent],
+    imports: [CommonModule],
+    exports: [LotProfileComponent],
+})
+export class GeneralModule {}
