+++ src/Auctionify.API/ClientApp/src/app/components/general/lot-profile/lot-profile.component.scss
@@ -10,7 +10,7 @@
     display: flex;
     flex-direction: column;
     gap: 48px;
-    
+
     &__photos-div {
         display: flex;
         gap: 8px;
@@ -25,7 +25,7 @@
         border-radius: 16px;
     }
 
-    &__main-photo img{
+    &__main-photo img {
         border-radius: 16px;
     }
 
@@ -73,7 +73,7 @@
     }
 
     &__wishlist ul li {
-        
+
         color: #2B5293;
     }
 
@@ -101,11 +101,11 @@
         padding-left: 24px;
         display: flex;
         gap: 80px;
-    } 
-    
+    }
+
     &__bids-info p {
         padding-top: 10px;
-    } 
+    }
 
     &__bid-btns {
         display: flex;
