+++ src/Auctionify.API/ClientApp/src/app/components/seller/seller-routing.module.ts
@@ -2,7 +2,6 @@ import { RouterModule, Routes } from '@angular/router';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { CreateLotComponent } from './create-lot/create-lot.component';
 import { NgModule } from '@angular/core';
-import { LotProfileSellerComponent } from './lot-profile-seller/lot-profile-seller.component';
 
 const sellerRoutes: Routes = [
     {
@@ -24,11 +23,6 @@ const sellerRoutes: Routes = [
         component: CreateLotComponent,
         pathMatch: 'full',
     },
-    {
-        path: 'lot-profile/:id',
-        component: LotProfileSellerComponent,
-        pathMatch: 'full',
-    }
 ];
 
 @NgModule({
